Refactor: add kaiko endpoint folder structure
