+++ packages/sources/kaiko/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as trades from './trades'
