+++ packages/sources/kaiko/src/index.ts
@@ -1,6 +1,7 @@
 import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute } from './adapter'
+import { makeExecute, endpointSelector } from './adapter'
 import { makeConfig, NAME } from './config'
+import * as types from './endpoint'
 
-const { server } = expose(NAME, makeExecute())
-export { NAME, makeExecute, makeConfig, server }
+const { server } = expose(NAME, makeExecute(), undefined, endpointSelector)
+export { NAME, makeExecute, makeConfig, server, types }
